changed chi text based on element

+++ src/main/java/com/crowsofwar/avatar/common/bending/BendingStyle.java
@@ -124,7 +124,7 @@ public abstract class BendingStyle implements ReadableWritable {
 	 */
 	public abstract String getName();
 
-	public abstract double getTextColour();
+	public abstract int getTextColour();
 
 	public abstract UUID getId();
